+++ packages/ui-shared/src/components/Table/util/index.ts
@@ -1,0 +1,2 @@
+export * from './columns';
+export * from './diagnosisColor';
